Eliminate deprecated content
