--- build_projects/Microsoft.DotNet.Cli.Build.Framework/UndefinedTargetException.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Microsoft.DotNet.Cli.Build.Framework
-{
-    public class UndefinedTargetException : Exception 
-    { 
-        public UndefinedTargetException(string message) : base(message) { }
-    }
-}
