+++ build_projects/shared-build-targets-utils/BuildFailureException.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace Microsoft.DotNet.Cli.Build
+{
+    public class BuildFailureException : Exception
+    {
+        public BuildFailureException()
+        {
+        }
+
+        public BuildFailureException(string message) : base(message)
+        {
+        }
+        
+        public BuildFailureException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+    }
+}
