+++ build_projects/shared-build-targets-utils/Utils/Utils.cs
@@ -90,18 +90,5 @@ namespace Microsoft.DotNet.Cli.Build
         {
             return $@"{commitHash}{Environment.NewLine}{version}{Environment.NewLine}";
         }
-
-        public static string GetSharedFrameworkVersionFileContent(BuildTargetContext c)
-        {
-            string SharedFrameworkNugetVersion = c.BuildContext.Get<string>("SharedFrameworkNugetVersion");
-            return $@"{c.BuildContext["CommitHash"]}{Environment.NewLine}{SharedFrameworkNugetVersion}{Environment.NewLine}";
-        }
-
-        public static string GetCliVersionFileContent(BuildTargetContext c)
-        {
-            var buildVersion = c.BuildContext.Get<BuildVersion>("BuildVersion");
-            var version = buildVersion.NuGetVersion;
-            return $@"{c.BuildContext["CommitHash"]}{Environment.NewLine}{version}{Environment.NewLine}";
-        }
     }
 }
